// packages/client/src/services/resident/model.ts
export interface Resident {
  id: number;
  image: string;
  ho_ten: string;
  age: number;
  shsq: string;
  nam_sinh: string
  que_quan: string;
  tru_quan: string;
  dia_chi: string;
  van_hoa: string;
  dan_toc: string;
  doan_vien: boolean;
  dang: boolean;
  ngay_vao_dang: string;
  ton_giao: string;
  suc_khoe: number;
  nghe_nghiep:string;
  noi_lam_viec:string;
  so_dien_thoai: string;
  ho_ten_cha: string;
  ho_ten_me: string;
  ho_ten_vo: string;
  so_con: number;
  thanh_phan_xuat_than: string;
  nhap_ngu: string;
  xuat_ngu: string;
  ten_don_vi: string;
  cap_bac: string;
  chuc_vu: string;
  ghi_chu: string;
}

export interface Resident1 {
  id: number;
  image: string;
  ho_ten: string;
  age: number;
  shsq: string;
  nam_sinh: string
  que_quan: string;
  tru_quan: string;
  dia_chi: string;
  van_hoa: string;
  dan_toc: string;
  doan_vien: boolean;
  dang: boolean;
  ngay_vao_dang: string;
  ton_giao: string;
  suc_khoe: number;
  nghe_nghiep:string;
  noi_lam_viec:string;
  so_dien_thoai: string;
  ho_ten_cha: string;
  ho_ten_me: string;
  ho_ten_vo: string;
  so_con: number;
  thanh_phan_xuat_than: string;
  nhap_ngu: string;
  xuat_ngu: string;
  ten_don_vi: string;
  cap_bac: string;
  chuc_vu: string;
  ghi_chu: string;
}

export interface Resident2 {
  id: number;
  image: string;
  ho_ten: string;
  age: number;
  shsq: string;
  nam_sinh: string
  que_quan: string;
  tru_quan: string;
  dia_chi: string;
  van_hoa: string;
  dan_toc: string;
  doan_vien: boolean;
  dang: boolean;
  ngay_vao_dang: string;
  ton_giao: string;
  suc_khoe: number;
  nghe_nghiep:string;
  noi_lam_viec:string;
  so_dien_thoai: string;
  ho_ten_cha: string;
  ho_ten_me: string;
  ho_ten_vo: string;
  so_con: number;
  thanh_phan_xuat_than: string;
  nhap_ngu: string;
  xuat_ngu: string;
  ten_don_vi: string;
  cap_bac: string;
  chuc_vu: string;
  ghi_chu: string;
}